List Using Liked-Based Implementation
A linked list is a linear data structure, in which the elements are not stored at contiguous memory locations. The elements in a linked list are linked using pointers 
In simple words, a linked list consists of nodes where each node contains a data field and a reference(link) to the next node in the list.
So, it has an unlimited size.